Move file loading to dedicated file
