+++ src/main.cpp
@@ -1,19 +1,14 @@
 #include "polyscope/polyscope.h"
-
-#include <igl/readOBJ.h>
-#include <igl/per_vertex_normals.h>
-
 #include "polyscope/point_cloud.h"
 
 #include "polyscopeSlicer.hpp"
-
-#include "ImGuiFileDialog.h"
 
 #include <iostream>
 #include <utility>
 
 // This file defines all the main types + data shared across the application components
 #include "./context.hpp"
+#include "./io.hpp"
 
 Context context;
 
@@ -246,118 +241,12 @@ Scalar evalScalarField_impl(const VectorType& input_pos)
     return current_value;
 }
 
-
-bool loadFile(const std::string& path)
-{
-    std::cout << "[Poncascope] Load file " << path << std::endl;
-    // first block: output cloudV and cloudN
-    Eigen::MatrixXd newCloud, newNormals;
-    {
-        bool worked;
-        Eigen::MatrixXi meshF;
-        measureTime( "[libIGL] obj file loading", [path, &newCloud, &meshF, &worked]()
-        // For convenience: use libIGL to load a mesh, and store only the vertices location and normal vector
-        {
-            const std::string filename = path.c_str();
-            worked = igl::readOBJ(filename, newCloud, meshF);
-        } );
-
-        if (worked) {
-            if (meshF.cols()==3) // we have a triangle mesh
-            {
-                igl::per_vertex_normals(newCloud, meshF, newNormals);
-            }
-        }
-        else {
-            std::cerr << "[libIGL] An error occurred when loading file " << path
-                      << std::endl;
-            return false;
-        }
-    }
-
-    // Check if normals have been properly loaded
-    /// \fixme : should not abort, but rather compute normals using Ponca.
-    {
-        int nbUnitNormal = int(newNormals.rowwise().squaredNorm().sum());
-        if ( nbUnitNormal != newCloud.rows() ) {
-            std::cerr << "[Poncascope] Point cloud has no normals, aborting" << std::endl;
-            return false;
-        }
-    }
-
-    context.cloudV = newCloud;
-    // no need to delete the previous cloud, polyscope handles it
-    context.cloud = polyscope::registerPointCloud("cloud", context.cloudV);
-    context.cloudN = newNormals;
-
-    // Bounding Box (used in the slicer)
-    context.lower = context.cloudV.colwise().minCoeff();
-    context.upper = context.cloudV.colwise().maxCoeff();
-
-    // Build Ponca KdTree
-    measureTime( "[Ponca] Build KdTree", []() {
-        buildKdTree(context.cloudV, context.cloudN, context.tree);
-    });
-
-
-    // Compute default point and neighborhood size according to the mean density
-    measureTime( "[Ponca] Compute point radius according to mean knn distance", []() {
-        Scalar cloudMDist = 0;
-        constexpr Scalar pointSizeFactor = 0.25;
-        constexpr Scalar scaleFactor = 5;
-#pragma omp parallel for
-        for (int i = 0; i < context.tree.samples().size(); ++i)
-        {
-            Scalar pointMDist = 0;
-            VectorType p = context.tree.points()[i].pos();
-            ::internal::doOnKNeighbors(i, [&pointMDist,p](auto&& neighborhood){
-                for (int j : neighborhood){
-                    pointMDist += (p-context.tree.points()[j].pos()).norm();
-                }
-            });
-#pragma omp critical
-            cloudMDist += pointMDist/Scalar(context.kNN);
-        }
-        context.pointRadius = pointSizeFactor * cloudMDist/Scalar(context.tree.samples().size());
-        context.NSize = scaleFactor * cloudMDist/Scalar(context.tree.samples().size());
-    });
-
-
-
-    // Register the point cloud with Polyscope
-    context.cloud->setPointRadius(context.pointRadius);
-    polyscope::requestRedraw();
-
-    context.useKnnGraph = false;
-
-    std::cout << "[Poncascope] Loading file succeeded"<< std::endl;
-
-    return true;
-}
-
 /// Define Polyscope callbacks
 void callback() {
 
     ImGui::PushItemWidth(100);
 
-    // open Dialog Simple
-    if (ImGui::Button("Open File Dialog")) {
-        IGFD::FileDialogConfig config;
-        config.path = context.lastPath;
-        ImGuiFileDialog::Instance()->OpenDialog("ChooseFileDlgKey", "Choose File", ".obj", config);
-    }
-    // display
-    if (ImGuiFileDialog::Instance()->Display("ChooseFileDlgKey")) {
-        if (ImGuiFileDialog::Instance()->IsOk()) { // action if OK
-            std::string filePathName = ImGuiFileDialog::Instance()->GetFilePathName();
-            // std::string filePath = ImGuiFileDialog::Instance()->GetCurrentPath();
-
-            if (loadFile(filePathName)) context.lastPath = filePathName;
-        }
-
-        // close
-        ImGuiFileDialog::Instance()->Close();
-    }
+    callback_io(context);
 
     ImGui::Text("Acceleration Structure");
     bool knnGraphUIChanged = ImGui::Checkbox("Use KnnGraph", &context.useKnnGraph);
@@ -452,10 +341,10 @@ int main(int argc, char** argv) {
 
     if (argc > 1)
     {
-        loadFile(argv[1]);
+        loadFile(argv[1],context);
     }
     else
-        loadFile("assets/armadillo.obj");
+        loadFile("assets/armadillo.obj", context);
 
     // Add the callback
     polyscope::state::userCallback = callback;
